refactor: move simple cli use to new file
